requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
groq==0.9.0
langchain==0.2.6
langchain-groq==0.1.6
langchain-community==0.2.6
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.2.0
pydantic==2.7.4
streamlit==1.36.0
streamlit-extras==0.4.3
pandas==2.2.2
plotly==5.22.0
httpx==0.27.0
